Optimize `ZKillAura`
These 3 if statements only need to be evaluated once. Also, if `mc.world` was null, then `mc.world.getEntities()` would throw a NPE. This means that we could reasonably assume that it won't ever be null, as it's never come up as an issue in the past, and remove it, but that's not my call to make.